requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.6
h11==0.14.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.4
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
